requirements.txt
streamlit
google-generativeai
astrapy
numpy